pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eathy-xhs"
version = "0.1.0"
description = "Eathy App 小红书自动化内容发布系统"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "httpx>=0.25",
    "feedparser>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "mcp>=1.0",
]

[project.scripts]
eathy = "eathy.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/eathy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.0",
    "pytest-cov>=4.0",
]
